# Eliminate redundant cache lookup with range deletion