--- a/db/db_compaction_test.cc
+++ b/db/db_compaction_test.cc
@@ -305,9 +305,9 @@ TEST_F(DBCompactionTest, TestTableReaderForCompaction) {
   dbfull()->TEST_WaitForCompact();
   // Preloading iterator issues one table cache lookup and creates
   // a new table reader. One file is created for flush and one for compaction.
-  // Compaction inputs make no table cache look-up for data iterators and one
-  // look-up per compaction input file (three).
-  ASSERT_EQ(num_table_cache_lookup, 5);
+  // Compaction inputs make no table cache look-up for data/range deletion
+  // iterators
+  ASSERT_EQ(num_table_cache_lookup, 2);
   // Create new iterator for:
   // (1) 1 for verifying flush results
   // (2) 3 for compaction input files
@@ -327,9 +327,9 @@ TEST_F(DBCompactionTest, TestTableReaderForCompaction) {
   cro.target_level = 2;
   cro.bottommost_level_compaction = BottommostLevelCompaction::kForce;
   db_->CompactRange(cro, nullptr, nullptr);
-  // Only verifying compaction outputs issues two table cache lookup
-  // (one for data block, one for range deletion block).
-  ASSERT_EQ(num_table_cache_lookup, 2);
+  // Only verifying compaction outputs issues one table cache lookup
+  // for both data block and range deletion block).
+  ASSERT_EQ(num_table_cache_lookup, 1);
   // One for compaction input, one for verifying compaction results.
   ASSERT_EQ(num_new_table_reader, 2);