--- a/db/table_cache.h
+++ b/db/table_cache.h
@@ -58,16 +58,6 @@ class TableCache {
       HistogramImpl* file_read_hist = nullptr, bool for_compaction = false,
       Arena* arena = nullptr, bool skip_filters = false, int level = -1);
 
-  // Return an iterator over the range deletion meta-block for the specified
-  // file number.
-  // @param skip_filters Disables loading/accessing the filter block
-  // @param level The level this table is at, -1 for "not set / don't know"
-  InternalIterator* NewRangeDeletionIterator(const ReadOptions& options,
-                                             const InternalKeyComparator& icmp,
-                                             const FileDescriptor& fd,
-                                             HistogramImpl* file_read_hist,
-                                             bool skip_filters, int level);
-
   // If a seek to internal key "k" in specified file finds an entry,
   // call (*handle_result)(arg, found_key, found_value) repeatedly until
   // it returns false.